mititools/requirements.txt
pandas==1.3.4
PyYAML==6.0
Jinja2==3.0.1
frictionless==4.23.0